% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finalheight_diff.R
\name{calc_finalheight_diff}
\alias{calc_finalheight_diff}
\title{Calculate final plant height into new column}
\usage{
calc_finalheight_diff(
  data = ecotox,
  HEIGHT_F = ecotox$HEIGHT_F,
  HEIGHT_I = ecotox$HEIGHT_I
)
}
\arguments{
\item{HEIGHT_F}{final height of plant in cm (numeric, no default)}

\item{HEIGHT_I}{initial height of plant in cm (numeric, no default)}

\item{ecotox}{name of tibble 'ecotox' (tibble, no default)}
}
\value{
difference (numeric)
}
\description{
Calculate final plant height into new column
}
\examples{
calc_finalheight_diff (HEIGHT_F-HEIGHT_I)

}
