% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rootshoot_ratio.R
\name{calc_rootshoot_ratio}
\alias{calc_rootshoot_ratio}
\title{Calculate root/shoot ratio into new column}
\usage{
calc_rootshoot_ratio(
  data = ecotox,
  BELOWGROUND = ecotox$BELOWGROUND,
  ABOVE_GROUND = ecotox$ABOVE_GROUND
)
}
\arguments{
\item{BELOWGROUND}{weight of BELOWGROUND biomass in grams (numeric, no default)}

\item{ABOVE_GROUND}{weight of the ABOVE_GROUND biomass in grams (numeric, no default)}

\item{ecotox}{name of tibble 'ecotox' (tibble, no default)}
}
\value{
ratio (numeric)
}
\description{
Calculate root/shoot ratio into new column
}
\examples{
calc_rootshoot_ratio (BELOWGROUND/ABOVE_GROUND)

}
