% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_density}
\alias{calc_shrub_density}
\title{calcuate shrub density}
\usage{
calc_shrub_density(width = 2, length = 4, height = 1)
}
\arguments{
\item{width}{(numberic, default 4)}

\item{length}{(numeric, default 2)}

\item{height}{(numeric, default 1)}

\item{mass}{(numeric)}

\item{dens}{(numeric)}
}
\value{
dens (numeric)
}
\description{
calcuate shrub density
}
