% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_mass}
\alias{calc_shrub_mass}
\title{calculate estimated shrub mass}
\usage{
calc_shrub_mass(volume)
}
\arguments{
\item{volume}{(numeric)}
}
\value{
Mass of the Shrub (numeric)
}
\description{
calculate estimated shrub mass
}
\examples{
calc_shrub_mass (3.5)
}
