% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_vol}
\alias{calc_shrub_vol}
\title{calculate volume of shrub}
\usage{
calc_shrub_vol(length = 1, width = 1, height = 1)
}
\arguments{
\item{length}{Length of the shrub (numeric, default 1)}

\item{width}{Width of the shrub (numeric, default 1)}

\item{height}{Height of the shrub (numeric, default 1)}
}
\value{
Volume of the shrub (numeric)
}
\description{
calculate volume of shrub
}
\examples{
calc_shrub_vol(length = 2, width = 3, height = 2)
}
