% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotox_lin_reg.R
\name{ecotox_lin_reg}
\alias{ecotox_lin_reg}
\title{Create visualization - linear regression}
\usage{
ecotox_lin_reg(
  data = ecotox,
  rs_ratio = ecotox$rs_ratio,
  fh_diff = ecotox$fh_diff,
  DOSE = ecotox$DOSE,
  lm = lm(rs_ratio ~ fh_diff, ecotox)
)
}
\arguments{
\item{rs_ratio}{ratio of root (belowground biomass) to shoot (above-ground biomass)(numeric, no default)}

\item{fh_diff}{heightchange of plant in cm (numeric, no default)}

\item{DOSE}{dose of ammonium in ppm (numeric, no default)}

\item{ecotox}{name of tibble 'ecotox' (tibble, no default)}
}
\value{
linear_regression (graph)
}
\description{
Create visualization - linear regression
}
